backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
ffmpeg-python==0.2.0
pysrt==1.1.2

edge-tts==6.1.9
gtts==2.3.2
openai==1.3.5
google-cloud-texttospeech==2.14.1
pillow==10.1.0
pyyaml==6.0.1
httpx==0.25.1
yt-dlp==2023.10.13
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
sqlalchemy==2.0.23
aiofiles==23.2.1
whisper==1.1.10

jinja2>=3.1.1